+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - FIX: added suppprt for  parameters in magics (run_notebooks and schedule_notebooks --- orbit.py)
 ### **Removed**
 - REMOVED: changes to team-script-launcher...filesystem is now always used
+- REMOVED: orbit profile support (use podsettings)
 
 ## **[1.2.0]**
 ### **Added**